Добавлен функционал

+++ main-service/main-server/src/main/java/ru/practicum/server/dto/EventRequestStatusUpdateDto.java
@@ -5,6 +5,6 @@ import ru.practicum.server.enums.EventRequestStatusEnum;
 
 @Data
 public class EventRequestStatusUpdateDto {
-    private int[] requestIds;
+    private Long[] requestIds;
     private EventRequestStatusEnum status;
 }
